requirements.txt
opencv-python==4.11.0.86
pillow==11.1.0